+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - py3 support
 ### Fixed
+- always use full module name in sys.modules
 ### Changed
 - check searchpath for null values